Add new charm configuration option
It is desirable to be able to insert a YAML string "as is" into the
filebeat configuration file. Unfortunately, there currently isn't a
charm configuration option which offers that ability. This change adds
such an option.

Builds upon the great work of @verterok who did most of the
implementation work which was then massaged into the current form.